refactor(front): renamed `space/Menu.vue` to `space/SpaeMenu.vue`
